+++ Lost-Project/Entity.h
@@ -9,9 +9,9 @@
 class Entity
 {
 public:
-	uint8_t id;
+	uint16_t id;
 	std::string name;
 	bool isDestroy;
-	Entity(uint8_t id, std::string name) : id(id), name(name), isDestroy(false) {}
+	Entity(uint16_t id, std::string name) : id(id), name(name), isDestroy(false) {}
 };
 #endif // !ENTITY_H
